+++ dev/setdict.py
@@ -579,3 +579,44 @@ def bench():
         odds.intersection(primes)
         odds.union(primes)
         """
+
+import ubelt as ub  # NOQA
+
+
+class RorUDictType(type):
+    cls = ub.UDict
+
+    @classmethod
+    def __ror__(mcls, obj):
+        return mcls.cls(obj)
+
+    @classmethod
+    def __or__(mcls, obj):
+        return mcls.cls(obj)
+
+
+class RorUDict(RorUDictType.cls, metaclass=RorUDictType):
+    pass
+
+try:
+    type({'10': 10} | RorUDict)
+except Exception as ex:
+    print(f'ex={ex}')
+
+try:
+    RorUDict | {'10': 10}
+except Exception as ex:
+    print(f'ex={ex}')
+
+
+a = {1: 10, 2: 20, 3: 30, 5: 50, 7: 70, 11: 110}
+b = {1: 11, 2: 21, 3: 31, 5: 51, 8: 81, 13: 131}
+t1 = (RorUDict | a)
+t2 = (a | RorUDict)
+assert t1 == t2
+assert t1 is not t2
+assert isinstance(t1, RorUDictType.cls)
+assert isinstance(t2, RorUDictType.cls)
+
+(RorUDict | a) & b
+(a | RorUDict) & b