+++ ubelt/util_download.py
@@ -136,7 +136,7 @@ def download(url, fpath=None, dpath=None, fname=None, appname=None,
         >>>     ub.download(url, hasher='sha512', hash_prefix='BAD_HASH')
     """
     from ubelt import ProgIter as Progress
-    from ubelt import ensure_app_cache_dir
+    from ubelt.util_path import Path
     import shutil
     import tempfile
     import hashlib
@@ -151,7 +151,7 @@ def download(url, fpath=None, dpath=None, fname=None, appname=None,
         raise ValueError('Cannot specify fpath with dpath or fname')
     if fpath is None:
         if dpath is None:
-            dpath = ensure_app_cache_dir(appname or 'ubelt')
+            dpath = Path.appdir(appname or 'ubelt').ensuredir()
         if fname is None:
             fname = basename(url)
         fpath = join(dpath, fname)
@@ -377,7 +377,7 @@ def grabdata(url, fpath=None, dpath=None, fname=None, redo=False,
         >>> fpath = ub.grabdata(url2, fname=fname, hash_prefix=prefix2)
         >>> assert json.loads(stamp_fpath.read_text())['hash'][0].startswith(prefix2)
     """
-    from ubelt.util_platform import ensure_app_cache_dir
+    from ubelt.util_path import Path
     from ubelt.util_cache import CacheStamp
     if appname and dpath:
         raise ValueError('Cannot specify appname with dpath')
@@ -386,7 +386,7 @@ def grabdata(url, fpath=None, dpath=None, fname=None, redo=False,
 
     if fpath is None:
         if dpath is None:
-            dpath = ensure_app_cache_dir(appname or 'ubelt')
+            dpath = Path.appdir(appname or 'ubelt').ensuredir()
         if fname is None:
             fname = basename(url)
         fpath = join(dpath, fname)